% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mukiau-package.r, R/mukiau.r
\docType{package}
\name{mukiau}
\alias{mukiau}
\alias{mukiau-package}
\title{Mukiau - Face the Dark}
\usage{
mukiau(spell = "Why are you using S-JIS?", wrap_width = 25,
  force_wrap = FALSE, size = 12, fontface = "bold", family = "sans",
  colour = "white", x = 0.1, y = 0.7, hjust = 0)
}
\arguments{
\item{spell}{text to overlay}

\item{wrap_width}{width of paragraph}

\item{force_wrap}{if TRUE, break lines with strictly fixed length of characters}

\item{...}{text style passed to \code{ggplot2::annotate()}}
}
\description{
Daibutsu bless you.

mukiau
}
\examples{
mukiau()

# wrap
mukiau(wrap_width = 10)

# break words
mukiau(wrap_width = 10, force_wrap = TRUE)

mukiau("\\\\u306a\\\\u305c\\\\u3042\\\\u306a\\\\u305f\\\\u306fCP932\\\\u306a\\\\u3069\\\\u3092\\\\u4f7f\\\\u3046\\\\u306e\\\\u3067\\\\u3059\\\\u304b\\\\uff1f", wrap_width = 10, force_wrap = TRUE)

}

