% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mukiau-package.r, R/mukiau.r
\docType{package}
\name{mukiau}
\alias{mukiau}
\alias{mukiau-package}
\title{Mukiau - Face Against the Dark}
\usage{
mukiau(spell = "Why are you using S-JIS?", wrap_width = 25,
  force_wrap = FALSE, size = 12, fontface = "bold", family = "sans",
  colour = "white", x = 0.1, y = 0.7, hjust = 0)
}
\arguments{
\item{spell}{text to overlay}

\item{wrap_width}{width of paragraph}

\item{force_wrap}{if TRUE, break lines with strictly fixed length of characters}

\item{...}{text style passed to \code{ggplot2::annotate()}}
}
\description{
Daibutsu bless you.

mukiau
}
\examples{
mukiau()

# wrap
mukiau(wrap_width = 10)

# break words
mukiau(wrap_width = 10, force_wrap = TRUE)

mukiau("\\x82\\xc8\\x82\\xba\\x82\\xa0\\x82\\xc8\\x82\\xbd\\x82\\xcd\\x43\\x50\\x39\\x33\\x32\\x82\\xc8\\x82\\xc7\\x82\\xf0\\x8e\\x67\\x82\\xa4\\x82\\xcc\\x82\\xc5\\x82\\xb7\\x82\\xa9\\x81\\x48", wrap_width = 10, force_wrap = TRUE)

}

